ADEN-8857 generalizing btl-wrapper, moving cheshire
